Pragma once :(
