+++ include/eventhandler.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <functional>
 #include <iostream>
 #include <vector>
