+++ include/game.h
@@ -1,18 +1,10 @@
 #pragma once
 
+#include "math.h"
 #include "SDL.h"
 #include "player.h"
 #include "renderer.h"
 #include "time.h"
-
-class GameTexture {
- public:
-  GameTexture(SDL_Texture* texture, const SDL_Rect& currentFrame, const int& FrameOffset)
-      : texture(texture), CurrentFrame(currentFrame), frameOffset(FrameOffset) {}
-  SDL_Texture* texture;
-  SDL_Rect CurrentFrame;
-  int frameOffset;
-};
 
 struct Game {
  public:
@@ -24,7 +16,7 @@ struct Game {
 
   void HandleSDLEvents(Time& Time, Player& Player);
 
-  GameTexture LoadTexture(const char* SpriteLocation, const Vector4& CurrentFrame, const int& FrameOffset, const RenderWindow& Window) {
+  GameTexture LoadTexture(const char* SpriteLocation, Vector4 CurrentFrame, int FrameOffset, RenderWindow Window) {
     SDL_Texture* texture = window.LoadTexture(SpriteLocation);
 
     SDL_Rect currentFrame;
