+++ include/math.h
@@ -1,3 +1,4 @@
+#pragma once
 struct Vector2 {
  public:
   Vector2() : x(0), y(0) {}
