+++ include/player.h
@@ -1,3 +1,5 @@
+#pragma once
+#include "utils.h"
 #include "entity.h"
 
 class Player : public Entity {
