+++ include/time.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <iostream>
 
 #include "renderer.h"
